+++ test/cases/html-plugin-auto-inject/webpack.config.js
@@ -4,14 +4,18 @@ import HtmlWebpackPlugin from 'html-webpack-plugin';
 module.exports = {
     entry: './index.js',
     plugins: [
-        new HtmlWebpackPlugin({ template: './index.ejs' }),
+        new HtmlWebpackPlugin({
+            template: './index.ejs',
+            minify: false,
+            scriptLoading: 'blocking',
+        }),
         new ConcatPlugin({
             name: 'file',
             fileName: '[name].[hash:20].js',
             filesToConcat: ['../../fixtures/a.js', '../../fixtures/b.js'],
             attributes: {
-                async: true
-            }
+                async: true,
+            },
         }),
     ],
     devtool: 'source-map',